*レジストリ関連をFileSystem.hに移動 *レジストリの値書き込みのバグ修正 *FileSystem.hにini(初期化)ファイル関連を追加
